reafactor(ts): Select
