+++ source/components/Select/OptGroup.tsx
@@ -1,15 +1,22 @@
-import React from 'react';
+import * as React from 'react';
 import PropTypes from 'prop-types';
 import classNames from 'classnames';
 
-export default class OptGroup extends React.Component {
+interface OptGroupProps {
+  _isShow?: boolean;
+  children?: React.ReactNode | React.ReactChildren;
+  label?: React.ReactNode | string;
+  prefixCls?: string;
+}
+
+export default class OptGroup extends React.Component<OptGroupProps> {
   static isSelectOptGroup = true;
 
   static propTypes = {
     _isShow: PropTypes.bool, // INTERNAL USE ONLY
     children: PropTypes.node.isRequired,
     label: PropTypes.oneOfType([PropTypes.node, PropTypes.string]).isRequired,
-    prefixCls: PropTypes.string,
+    prefixCls: PropTypes.string
   };
 
   static defaultProps = {
@@ -17,18 +24,19 @@ export default class OptGroup extends React.Component {
     prefixCls: 'fishd-select-dropdown-option-group'
   };
 
-  constructor(props) {
+  constructor(props: OptGroupProps) {
     super(props);
   }
 
   render() {
-    const {children, label, prefixCls, _isShow} = this.props;
-    return _isShow &&
-      (
+    const { children, label, prefixCls, _isShow } = this.props;
+    return (
+      _isShow && (
         <div className={classNames(`${prefixCls}`)}>
           <p className={`${prefixCls}-label`}>{label}</p>
           {children}
         </div>
-      );
+      )
+    );
   }
 }
