+++ source/components/Select/index.tsx
@@ -3,9 +3,5 @@ import Option from './Option';
 import OptGroup from './OptGroup';
 import './style/index.less';
 
-Select.Option = Option;
-
-Select.OptGroup = OptGroup;
-
-export {Option, OptGroup};
+export { Option, OptGroup };
 export default Select;
